build_jwt/requirements.txt
PyJWT==2.4.0
cryptography==37.0.2